pyproject.toml
[project]
name = "toyshop"
version = "0.1.0"
description = "ToyShop - Software factory with OpenSpec contracts and OpenHands LLM"
requires-python = ">=3.12"
dependencies = ["openhands-sdk"]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["toyshop*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
